Revert "feat: Update to Docling 2.28 (#106)"
This reverts commit 20ec87a63a99145bc0ad7931549af8a0c30db641.
