+++ docling_serve/engines/base_orchestrator.py
@@ -1,6 +1,6 @@
 from abc import ABC, abstractmethod
 
-from docling_serve.datamodel.task import Task
+from docling_serve.datamodel.engines import Task
 
 
 class BaseOrchestrator(ABC):
